+++ Vezbe04/jpa-example/src/main/java/rs/ac/uns/ftn/informatika/jpa/dto/CourseDTO.java
@@ -3,7 +3,7 @@ package rs.ac.uns.ftn.informatika.jpa.dto;
 import rs.ac.uns.ftn.informatika.jpa.model.Course;
 
 public class CourseDTO {
-	private Long id;
+	private Integer id;
 	private String name;
 
 	public CourseDTO() {
@@ -14,13 +14,13 @@ public class CourseDTO {
 		this(course.getId(), course.getName());
 	}
 
-	public CourseDTO(Long id, String name) {
+	public CourseDTO(Integer id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
