+++ Vezbe04/jpa-example/src/main/java/rs/ac/uns/ftn/informatika/jpa/repository/CourseRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import rs.ac.uns.ftn.informatika.jpa.model.Course;
 
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 	
 	@Query("select c from Course c join fetch c.exams e where c.id =?1")
-	public Course findOneWithExams(Long courseId);
+	public Course findOneWithExams(Integer courseId);
 
 }
