+++ Vezbe04/jpa-example/src/main/java/rs/ac/uns/ftn/informatika/jpa/repository/ExamRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import rs.ac.uns.ftn.informatika.jpa.model.Exam;
 
-public interface ExamRepository extends JpaRepository<Exam, Long> {
+public interface ExamRepository extends JpaRepository<Exam, Integer> {
 
 }
