+++ Vezbe04/jpa-example/src/main/java/rs/ac/uns/ftn/informatika/jpa/repository/TeacherRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import rs.ac.uns.ftn.informatika.jpa.model.Teacher;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
 	
 	@Query("select t from Teacher t join fetch t.courses e where t.id =?1")
-	public Teacher findOneWithCourses(Long teacherId);
+	public Teacher findOneWithCourses(Integer teacherId);
 
 }
